f
f

f

fx

fx
